requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==1.5
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
